+++ internal/cli/show_test.go
@@ -118,7 +118,7 @@ func TestFormatTaskShowsDepends(t *testing.T) {
 func TestShowCommand(t *testing.T) {
 	dir := t.TempDir()
 	chdir(t, dir)
-	if err := runRoot(t, "init", "--template", "hierarchy"); err != nil {
+	if err := runRoot(t, "init", "--template", hierarchyTemplate(t)); err != nil {
 		t.Fatalf("init: %v", err)
 	}
 	if _, _, err := runOut(t, "add", "--type", "epic", "fix login"); err != nil {